requirements.txt
Flask==2.0.1
flask-cors==3.0.10
Pillow==8.3.2
numpy==1.21.2
opencv-python==4.5.3.56
python-base64==0.1.0
requests==2.26.0
